requirements.txt
aiogram
python-dotenv
alembic
sqlalchemy
asyncpg
psycopg2-binary
mistralai
pytz
